Widen utils_string_to_int's return to uint32_t
uint16_t truncated any parsed value over 65535 -- silently wrong for
anything sized like an HTTP Content-Length (image sizes routinely
exceed 64KB). All existing callers already narrow-cast the result to
smaller types, so this only fixes previously-incorrect large values.

+++ utils_string.c
@@ -33,9 +33,9 @@ bool utils_string_is_receive_data(char* buffer, uint16_t buffer_len, const char*
  * @param buffer_length Buffer Len
  * @return true Interger
  */
-uint16_t utils_string_to_int(char* buffer, uint16_t buffer_length)
+uint32_t utils_string_to_int(char* buffer, uint16_t buffer_length)
 {
-	uint16_t result = 0;
+	uint32_t result = 0;
 	for(uint16_t var = 0; var < buffer_length; ++var)
 	{
 		if((buffer[var] >= '0') && (buffer[var] <= '9'))
